Clean everything up and move to mustache.
