+++ assemble/src/main/assemblies/component.xml
@@ -40,47 +40,49 @@
         <include>org.apache.commons:commons-math3</include>
         <include>org.apache.commons:commons-vfs2</include>
         <include>org.apache.thrift:libthrift</include>
+        <include>org.eclipse.jetty:jetty-continuation</include>
         <include>org.eclipse.jetty:jetty-http</include>
         <include>org.eclipse.jetty:jetty-io</include>
         <include>org.eclipse.jetty:jetty-security</include>
         <include>org.eclipse.jetty:jetty-server</include>
+        <!-- jetty-servlet only needed by old monitor -->
         <include>org.eclipse.jetty:jetty-servlet</include>
         <include>org.eclipse.jetty:jetty-util</include>
+        <include>org.eclipse.jetty:jetty-continuation</include>
         <include>org.apache.htrace:htrace-core</include>
         <include>org.slf4j:slf4j-api</include>
         <include>org.slf4j:slf4j-log4j12</include>
-        <include>org.glassfish.grizzly:grizzly-framework</include>
-        <include>org.glassfish.grizzly:grizzly-http-server</include>
-        <include>org.glassfish.grizzly:grizzly-http</include>
-        <include>org.glassfish.jersey.containers:jersey-container-grizzly2-http</include>
-        <include>org.glassfish.jersey.containers:jersey-container-servlet-core</include>
-        <include>org.glassfish.jersey.ext:jersey-entity-filtering</include>
+        <!-- Jersey/Jackson-based webservice -->
         <include>com.fasterxml.jackson.core:jackson-annotations</include>
         <include>com.fasterxml.jackson.core:jackson-core</include>
         <include>com.fasterxml.jackson.core:jackson-databind</include>
         <include>com.fasterxml.jackson.jaxrs:jackson-jaxrs-base</include>
         <include>com.fasterxml.jackson.jaxrs:jackson-jaxrs-json-provider</include>
         <include>com.fasterxml.jackson.module:jackson-module-jaxb-annotations</include>
-        <include>org.glassfish.hk2.external:javax.inject</include>
+        <include>com.github.spullara.mustache.java:compiler</include>
         <include>javax.annotation:javax.annotation-api</include>
-        <include>javax.servlet:javax.servlet-api</include>
+        <include>javax.inject:javax.inject</include>
         <include>javax.validation:validation-api</include>
         <include>javax.ws.rs:javax.ws.rs-api</include>
+        <include>org.glassfish.hk2.external:aopalliance-repackaged</include>
+        <include>org.glassfish.hk2.external:javax.inject</include>
         <include>org.glassfish.hk2:hk2-api</include>
         <include>org.glassfish.hk2:hk2-locator</include>
         <include>org.glassfish.hk2:hk2-utils</include>
-        <include>org.glassfish.jersey.core:jersey-client</include>
+        <include>org.glassfish.hk2:osgi-resource-locator</include>
         <include>org.glassfish.jersey.bundles.repackaged:jersey-guava</include>
+        <include>org.glassfish.jersey.containers:jersey-container-jetty-http</include>
+        <include>org.glassfish.jersey.containers:jersey-container-servlet-core</include>
+        <include>org.glassfish.jersey.containers:jersey-container-servlet</include>
+        <include>org.glassfish.jersey.core:jersey-client</include>
         <include>org.glassfish.jersey.core:jersey-common</include>
-        <include>org.glassfish.jersey.media:jersey-media-json-jackson</include>
         <include>org.glassfish.jersey.core:jersey-server</include>
-        <!--include>javax.inject:javax.inject</include-->
-        <!--include>javax.servlet.jsp:jsp-api</include-->
-        <!--include>javax.servlet:servlet-api</include-->
-        <!--include>org.glassfish:javax.servlet</include-->
-        <!--include>org.glassfish.grizzly:grizzly-http-servlet</include-->
-        <!--include>org.glassfish.hk2.external:aopalliance-repackaged</include-->
-        <!--include>org.glassfish.hk2:osgi-resource-locator</include-->
+        <include>org.glassfish.jersey.ext:jersey-entity-filtering</include>
+        <include>org.glassfish.jersey.ext:jersey-mvc-mustache</include>
+        <include>org.glassfish.jersey.ext:jersey-mvc</include>
+        <include>org.glassfish.jersey.media:jersey-media-jaxb</include>
+        <include>org.glassfish.jersey.media:jersey-media-json-jackson</include>
+        <include>org.javassist:javassist</include>
       </includes>
       <excludes>
         <exclude>${groupId}:${artifactId}-docs</exclude>
