+++ server/monitor/src/main/resources/templates/org/apache/accumulo/monitor/rest/view/Overview/index.mustache
@@ -17,8 +17,8 @@
 <html>
 <body>
     <h2>Jersey RESTful Web Application!</h2>
-    <p><a href="webapi/myresource">Jersey resource</a>
     <p>Visit <a href="http://jersey.java.net">Project Jersey website</a>
     for more information on Jersey!
+    <h3>foo = {{ foo }}</h3>
 </body>
 </html>
